Added a test for single_data_transfer

+++ src/defns.h
@@ -1,12 +1,14 @@
 #ifndef DEFN
 #define DEFN
 
+#include<stdint.h>
+
 /*
  * Type Definitions for types used across project
  */
 
 typedef unsigned char byte_t;
-typedef unsigned long int word_t;
+typedef uint32_t word_t;
 typedef unsigned char nibble_t;
 typedef unsigned char bit_t;
